# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateEntityRequest:
    properties:
      address:
        type: string
      banner_url:
        type: string
      category:
        type: string
      city:
        type: string
      description:
        type: string
      gallery:
        description: List of Media IDs (UUIDs)
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      profile_url:
        type: string
    required:
    - name
    type: object
...
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Entity:
    properties:
      address:
        type: string
      banner_url:
        type: string
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        description: e.g., "Food", "Services"
        type: string
      city:
        type: string
      contact_info:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_verified:
        description: Check dorado
        type: boolean
      latitude:
        description: |-
          PostGIS Geography Point (SRID 4326)
          PostGIS Geography Point (SRID 4326)
          Location           string             `gorm:"type:geography(POINT,4326)" json:"-"`
        type: number
      longitude:
        type: number
      name:
        type: string
      owner_id:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.EntityPhoto'
        type: array
      profile_url:
        type: string
      updated_at:
        type: string
      verification_status:
        $ref: '#/definitions/models.VerificationStatus'
    type: object
  models.EntityPhoto:
    properties:
      entity_id:
        type: string
      id:
        type: string
      media:
        allOf:
        - $ref: '#/definitions/models.Media'
        description: Associations
      media_id:
        type: string
      order:
        type: integer
    type: object
  models.Media:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      id:
        type: string
      original_name:
        type: string
      size:
        type: integer
    type: object
...
  models.VerificationStatus:
    enum:
    - pending
    - verified
    - rejected
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusVerified
    - StatusRejected
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Interactive API documentation for the Local Discovery App.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Empre Backend API
  version: "1.0"
paths:
...
  /api/entities:
    get:
      description: Search entities with geographic and category filters
      parameters:
      - description: Latitude
        in: query
        name: lat
        type: number
      - description: Longitude
        in: query
        name: long
        type: number
      - description: Radius in meters
        in: query
        name: radius
        type: number
      - description: Category UUID
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Entity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Find all entities
      tags:
      - Entities
    post:
      consumes:
      - application/json
      description: Register a new business entity with basic details
      parameters:
      - description: Entity Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateEntityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Entity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new entity
      tags:
      - Entities
  /api/entities/{id}:
    delete:
      description: Remove a specific business entity (Owner only)
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete entity
      tags:
      - Entities
    get:
      description: Get full details of a specific business entity
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Find entity by ID
      tags:
      - Entities
    put:
      consumes:
      - application/json
      description: Update details of a specific business entity (Owner only)
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update entity
      tags:
      - Entities
  /api/entities/{id}/images:
    post:
      consumes:
      - multipart/form-data
      description: Upload a profile, banner, or gallery image for a specific entity
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Image File
        in: formData
        name: file
        required: true
        type: file
      - description: Image Type (profile, banner, gallery)
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload entity image
      tags:
      - Entities
  /api/entities/mine:
    get:
      description: Get all business entities owned by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Entity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Find my entities
      tags:
      - Entities
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"